requirements.txt
fastapi
psycopg2-binary
pydantic_settings
python-dotenv
pydantic
schedule
requests
sqlalchemy
alembic
uvicorn